egcs gave a warning in function offset; fixed it

+++ code/tables/implement/Tables/ColumnCache.h
@@ -124,7 +124,8 @@ inline void ColumnCache::invalidate()
 
 inline Int ColumnCache::offset (uInt rownr) const
 {
-    return rownr<itsStart || rownr>itsEnd  ?  -1 : (rownr-itsStart)*itsIncr;
+    return rownr<itsStart || rownr>itsEnd  ?  -1 :
+	                                      Int((rownr-itsStart)*itsIncr);
 }
 
 inline const void* ColumnCache::dataPtr() const
